Fix thaliMobileNativeWrapper tests, move executeZombieProofTest to thaliMobileNativeTestUtils, move getSamePeerWithRetry from testUtils to thaliMobileNativeTestUtils
